+++ oc-includes/osclass/classes/datatables/ItemsDataTable.php
@@ -392,12 +392,11 @@ class ItemsDataTable extends DataTable
                 }
                 $row['category']   = osc_esc_html($aRow['s_category_name']);
                 $row['location']   = $this->get_row_location();
-                $row['date']       = osc_format_date($aRow['dt_pub_date'], osc_date_format() . ' ' . osc_time_format());
+                $row['date']       = osc_admin_date($aRow['dt_pub_date']);
                 $row['expiration'] =
-                    ($aRow['dt_expiration'] !== '9999-12-31 23:59:59') ? osc_format_date(
-                        $aRow['dt_expiration'],
-                        osc_date_format() . ' ' . osc_time_format()
-                    ) : __('Never expires');
+                    ($aRow['dt_expiration'] !== '9999-12-31 23:59:59')
+                        ? osc_admin_date($aRow['dt_expiration'])
+                        : __('Never expires');
 
                 $row = osc_apply_filter('items_processing_row', $row, $aRow);
 
@@ -773,13 +772,11 @@ class ItemsDataTable extends DataTable
                 $row['rep']         = $aRow['i_num_repeated'];
                 $row['exp']         = $aRow['i_num_expired'];
                 $row['off']         = $aRow['i_num_offensive'];
-                $row['date']        =
-                    osc_format_date($aRow['dt_pub_date'], osc_date_format() . ' ' . osc_time_format());
+                $row['date']        = osc_admin_date($aRow['dt_pub_date']);
                 $row['expiration']  =
-                    ($aRow['dt_expiration'] !== '9999-12-31 23:59:59') ? osc_format_date(
-                        $aRow['dt_expiration'],
-                        osc_date_format() . ' ' . osc_time_format()
-                    ) : __('Never expires');
+                    ($aRow['dt_expiration'] !== '9999-12-31 23:59:59')
+                        ? osc_admin_date($aRow['dt_expiration'])
+                        : __('Never expires');
 
                 $row = osc_apply_filter('items_processing_reported_row', $row, $aRow);